+++ R/processing.R
@@ -666,6 +666,8 @@ get_args_man <- function(alg = "", options = TRUE,
 #'   [get_args_man()] for more details. Please note that you can either specify
 #'   R arguments directly via the triple dots (see above) or via the
 #'   parameter-argument list. However, you may not mix the two methods.
+#' @param NA_flag Value used for NAs when exporting raster objects through
+#'   [save_spatial_objects()] (default: -99999).
 #' @param qgis_env Environment containing all the paths to run the QGIS API. For
 #'   more information, refer to [set_env()].
 #' @return The function returns the complete parameter-argument list for a given
@@ -728,7 +730,7 @@ get_args_man <- function(alg = "", options = TRUE,
 #' pass_args(alg, elevation = dem, format = "degrees")
 #' }
 
-pass_args <- function(alg, ..., params = NULL, qgis_env = set_env()) {
+pass_args <- function(alg, ..., params = NULL, NA_flag = -99999, qgis_env = set_env()) {
   dots <- list(...)
   if (!is.null(params) && (length(dots) > 0)) {
     stop(paste(
@@ -852,7 +854,8 @@ pass_args <- function(alg, ..., params = NULL, qgis_env = set_env()) {
   # certainly not an output file)
   params[!out$output] <- save_spatial_objects(
     params = params[!out$output],
-    type_name = out$type_name
+    type_name = out$type_name,
+    NA_flag = NA_flag
   )
 
   # if the user has only specified an output filename without a directory path,
@@ -925,43 +928,44 @@ pass_args <- function(alg, ..., params = NULL, qgis_env = set_env()) {
   params
 }
 
-#' @title Interface to QGIS commands
-#' @description `run_qgis` calls QGIS algorithms from within R while passing the
-#'   corresponding function arguments.
-#' @param alg Name of the GIS function to be used (see [find_algorithms()]).
-#' @param ... Triple dots can be used to specify QGIS geoalgorithm arguments as
-#'   R named arguments. For more details, please refer to [pass_args()].
-#' @param params Parameter-argument list for a specific geoalgorithm. Please
-#'   note that you can either specify R named arguments directly via the triple
-#'   dots (see above) or via a parameter-argument list. However, you may not mix
-#'   the two methods. See the example section, [pass_args()] and
-#'   [get_args_man()] for more details.
-#' @param load_output If `TRUE`, all QGIS output files ([sf::sf()]-object in the
-#'   case of vector data and [raster::raster()]-object in the case of a raster)
-#'   specified by the user (i.e. the user has to indicate output files) will be
-#'   loaded into R. A list will be returned if there is more than one output
-#'   file (e.g., `grass7:r.slope.aspect`). See the example section for more
-#'   details.
-#' @param show_output_paths Logical. QGIS computes all possible output files for
-#'   a given geoalgorithm, and saves them to a temporary location in case the
-#'   user has not specified explicitly another output location. Setting
-#'   `show_output` to `TRUE` (the default) will print all output paths to the
-#'   console after the successful geoprocessing.
-#' @param qgis_env Environment containing all the paths to run the QGIS API. For
-#'   more information, refer to [set_env()].
-#' @details This workhorse function calls the QGIS Python API, and specifically
-#'   `processing.runalg`.
-#' @return The function prints a list (named according to the output parameters)
-#'   containing the paths to the files created by QGIS. If not otherwise
-#'   specified, the function saves the QGIS generated output files to a
-#'   temporary folder (created by QGIS). Optionally, function parameter
-#'   `load_output` loads spatial QGIS output (vector and raster data) into R.
-#' @note Please note that one can also pass spatial R objects as input
-#'   parameters where suitable (e.g., input layer, input raster). Supported
-#'   formats are [sp::SpatialPointsDataFrame()]-,
-#'   [sp::SpatialLinesDataFrame()]-, [sp::SpatialPolygonsDataFrame()]-,
-#'   [sf::sf()]- (of class `sf`, `sfc` as well as `sfg`), and
-#'   [raster::raster()]-objects. See the example section for more details.
+#'@title Interface to QGIS commands
+#'@description `run_qgis` calls QGIS algorithms from within R while passing the
+#'  corresponding function arguments.
+#'@param alg Name of the GIS function to be used (see [find_algorithms()]).
+#'@param ... Triple dots can be used to specify QGIS geoalgorithm arguments as R
+#'  named arguments. For more details, please refer to [pass_args()].
+#'@param params Parameter-argument list for a specific geoalgorithm. Please note
+#'  that you can either specify R named arguments directly via the triple dots
+#'  (see above) or via a parameter-argument list. However, you may not mix the
+#'  two methods. See the example section, [pass_args()] and [get_args_man()] for
+#'  more details.
+#'@param load_output If `TRUE`, all QGIS output files ([sf::sf()]-object in the
+#'  case of vector data and [raster::raster()]-object in the case of a raster)
+#'  specified by the user (i.e. the user has to indicate output files) will be
+#'  loaded into R. A list will be returned if there is more than one output file
+#'  (e.g., `grass7:r.slope.aspect`). See the example section for more details.
+#'@param show_output_paths Logical. QGIS computes all possible output files for
+#'  a given geoalgorithm, and saves them to a temporary location in case the
+#'  user has not specified explicitly another output location. Setting
+#'  `show_output` to `TRUE` (the default) will print all output paths to the
+#'  console after the successful geoprocessing.
+#'@param NA_flag Value used for NAs when exporting raster objects through
+#'  [pass_args()] and [save_spatial_objects()] (default: -99999).
+#'@param qgis_env Environment containing all the paths to run the QGIS API. For
+#'  more information, refer to [set_env()].
+#'@details This workhorse function calls the QGIS Python API, and specifically
+#'  `processing.runalg`.
+#'@return The function prints a list (named according to the output parameters)
+#'  containing the paths to the files created by QGIS. If not otherwise
+#'  specified, the function saves the QGIS generated output files to a temporary
+#'  folder (created by QGIS). Optionally, function parameter `load_output` loads
+#'  spatial QGIS output (vector and raster data) into R.
+#'@note Please note that one can also pass spatial R objects as input parameters
+#'  where suitable (e.g., input layer, input raster). Supported formats are
+#'  [sp::SpatialPointsDataFrame()]-, [sp::SpatialLinesDataFrame()]-,
+#'  [sp::SpatialPolygonsDataFrame()]-, [sf::sf()]- (of class `sf`, `sfc` as well
+#'  as `sfg`), and [raster::raster()]-objects. See the example section for more
+#'  details.
 #'
 #'  GRASS users do not have to specify manually the GRASS region extent
 #'  (function argument GRASS_REGION_PARAMETER). If "None" (the QGIS default),
@@ -969,13 +973,13 @@ pass_args <- function(alg, ..., params = NULL, qgis_env = set_env()) {
 #'  the region extent based on the user-specified input layers. If you do want
 #'  to specify it yourself, please do it in accordance with the [QGIS
 #'  documentation](https://docs.qgis.org/2.8/en/docs/user_manual/processing/console.html),
-#'  i.e., use a character string and separate the coordinates with a comma:
+#'   i.e., use a character string and separate the coordinates with a comma:
 #'  "xmin, xmax, ymin, ymax".
 #'
-#' @author Jannes Muenchow, Victor Olaya, QGIS core team
-#' @export
-#' @importFrom sf read_sf
-#' @importFrom raster raster
+#'@author Jannes Muenchow, Victor Olaya, QGIS core team
+#'@export
+#'@importFrom sf read_sf
+#'@importFrom raster raster
 #' @examples
 #' \dontrun{
 #' # calculate the slope of a DEM
@@ -1001,7 +1005,8 @@ pass_args <- function(alg, ..., params = NULL, qgis_env = set_env()) {
 #' }
 
 run_qgis <- function(alg = NULL, ..., params = NULL, load_output = FALSE,
-                     show_output_paths = TRUE, qgis_env = set_env()) {
+                     show_output_paths = TRUE, NA_flag = -99999,
+                     qgis_env = set_env()) {
 
   # check if the QGIS application has already been started
   tmp <- try(expr = open_app(qgis_env = qgis_env), silent = TRUE)
@@ -1030,7 +1035,8 @@ run_qgis <- function(alg = NULL, ..., params = NULL, load_output = FALSE,
 
 
   # construct a parameter-argument list using get_args_man and user input
-  params <- pass_args(alg, ..., params = params, qgis_env = qgis_env)
+  params <- pass_args(alg, ..., params = params, NA_flag = NA_flag,
+                      qgis_env = qgis_env)
 
   # build the Python command
   # r_to_py(params) would also create a dictionary which would be a rather